Move Windows-specific things into their own files
So that they are only compiled on Windows.
