+++ pkg/node/nodename_test.go
@@ -17,23 +17,15 @@ limitations under the License.
 package node
 
 import (
-	"context"
-	"net"
-	"net/http"
-	"net/url"
 	"runtime"
 	"testing"
 
-	"github.com/k0sproject/k0s/pkg/k0scontext"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	nodeutil "k8s.io/component-helpers/node/util"
 )
 
 func TestGetNodename(t *testing.T) {
-	kubeHostname, err := nodeutil.GetHostname("")
-	require.NoError(t, err)
-
 	t.Run("should_always_return_override_if_given", func(t *testing.T) {
 		name, err := GetNodename("override")
 		if assert.NoError(t, err) {
@@ -42,59 +34,14 @@ func TestGetNodename(t *testing.T) {
 	})
 
 	if runtime.GOOS != "windows" {
+		kubeHostname, err := nodeutil.GetHostname("")
+		require.NoError(t, err)
+
 		t.Run("should_call_kubernetes_hostname_helper_on_linux", func(t *testing.T) {
 			name, err := GetNodename("")
 			if assert.NoError(t, err) {
 				assert.Equal(t, kubeHostname, name)
 			}
 		})
-	} else {
-		baseURL := startFakeMetadataServer(t)
-
-		t.Run("windows_no_metadata_service_available", func(t *testing.T) {
-			ctx := k0scontext.WithValue(context.TODO(), nodenameURL(baseURL))
-			name, err := getNodename(ctx, "")
-			if assert.NoError(t, err) {
-				assert.Equal(t, kubeHostname, name)
-			}
-		})
-
-		t.Run("windows_metadata_service_is_available", func(t *testing.T) {
-			ctx := k0scontext.WithValue(context.TODO(), nodenameURL(baseURL+"/latest/meta-data/local-hostname"))
-			name, err := getNodename(ctx, "")
-			if assert.NoError(t, err) {
-				assert.Equal(t, "some-hostname from aws_metadata", name)
-			}
-		})
 	}
-}
-
-func startFakeMetadataServer(t *testing.T) string {
-	mux := http.NewServeMux()
-	mux.HandleFunc("/latest/meta-data/local-hostname", func(w http.ResponseWriter, r *http.Request) {
-		_, err := w.Write([]byte("Some-hostname from AWS_metadata\n"))
-		assert.NoError(t, err)
-	})
-	server := &http.Server{Addr: "localhost:0", Handler: mux}
-	listener, err := net.Listen("tcp", server.Addr)
-	if err != nil {
-		require.NoError(t, err)
-	}
-
-	serverError := make(chan error)
-	go func() {
-		defer close(serverError)
-		serverError <- server.Serve(listener)
-	}()
-
-	t.Cleanup(func() {
-		err := server.Shutdown(context.Background())
-		if !assert.NoError(t, err, "Couldn't shutdown HTTP server") {
-			return
-		}
-
-		assert.ErrorIs(t, <-serverError, http.ErrServerClosed, "HTTP server terminated unexpectedly")
-	})
-
-	return (&url.URL{Scheme: "http", Host: listener.Addr().String()}).String()
 }
